refactor(environmental): Consolidate duplicate viewset implementations
- Audit and consolidate duplicate implementations in views.py and api/viewsets.py
- Migrate missing functionality from views.py to api/viewsets.py:
  * Add by_container() action to EnvironmentalReadingViewSet
  * Add by_area() action to WeatherDataViewSet
- Retain all optimized functionality from api/viewsets.py:
  * recent() action with DISTINCT ON optimization
  * stats() action for aggregations
  * Full serializers with parameter bounds validation
  * Explicit authentication (TokenAuthentication, JWTAuthentication)
  * Query optimizations with select_related
- Delete deprecated files:
  * apps/environmental/views.py (261 lines removed)
  * apps/environmental/urls.py (23 lines removed)
- All 81 environmental tests passing with no regressions

Benefits:
- Single source of truth for environmental viewsets
- Eliminates code duplication and maintenance burden
- Retains all functionality from both implementations
- Better query performance with optimizations
- Consistent authentication and permission handling

Routing unchanged - uses apps/environmental/api/routers.py exclusively

Updates REMEDIATION_PLAN.md Task 21 as completed
